Ready

+++ assign/lloyd_assign/lloyd_assign.h
@@ -4,12 +4,13 @@
 #define LLOYDS_ASSIGN_H
 
 #include <iostream>
+#include <cmath>
 #include "../../clustering/cluster.h"
 
 class dataset;
 
 // returns index of min value
-int find_min(float* list, int k);
+int find_min(double* list, int k);
 dataset* lloyd(dataset* data, dataset** centers, int n, int k, std::string metric);
 
 #endif
